Create a separate CadPicker class
